heroes-chatbot-back/requirements.txt
fastapi
uvicorn
python-socketio[asyncio]
httpx
openai
python-dotenv
pandas
openpyxl
pykrx
